:art: Code Refactoring
Restructured directory
