+++ api/afdb/library.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { Subject, Grade, Chapter, Resource, Topic } from '../types'
+import { Subject, Grade, Chapter, Resource, Topic } from '../../app/types'
 
 const url = process.env.NEXT_PUBLIC_AF_DB_SERVICE_URL;
 const bearerToken = process.env.NEXT_PUBLIC_AF_DB_SERVICE_BEARER_TOKEN;
