+++ app/reports/page.tsx
@@ -5,7 +5,7 @@ import ReportsList from "./reports_list";
 import Loading from "../loading";
 import BottomNavigationBar from "@/components/BottomNavigationBar";
 import TopBar from "@/components/TopBar";
-import { useAuth } from "../AuthContext";
+import { useAuth } from "../../services/AuthContext";
 
 export default function ReportsPage() {
     const { loggedIn } = useAuth();
@@ -13,13 +13,13 @@ export default function ReportsPage() {
     return (
         <>
             {/* {loggedIn ? ( */}
-                <main className="max-w-xl mx-auto bg-white">
-                    <TopBar />
-                    <Client message="">
-                        <ReportsList />
-                    </Client>
-                    <BottomNavigationBar />
-                </main>
+            <main className="max-w-xl mx-auto bg-white">
+                <TopBar />
+                <Client message="">
+                    <ReportsList />
+                </Client>
+                <BottomNavigationBar />
+            </main>
             {/* ) : (
                 <main className="max-w-xl mx-auto bg-white">
                     <TopBar />
