+++ utils/dateUtils.ts
@@ -1,0 +1,15 @@
+export function isSameDay(date1: Date, date2: Date): boolean {
+    return (
+        date1.getDate() === date2.getDate() &&
+        date1.getMonth() === date2.getMonth() &&
+        date1.getFullYear() === date2.getFullYear()
+    );
+}
+
+
+export function formatTime(dateTimeStr: string) {
+    const date = new Date(dateTimeStr);
+    const hours = String(date.getUTCHours()).padStart(2, "0");
+    const minutes = String(date.getUTCMinutes()).padStart(2, "0");
+    return `${hours}:${minutes}`;
+}
